requirements.txt
aiogram==3.10.0
aiohttp==3.9.3
google-generativeai>=0.7.2